Added the Pygame version
It's a more visual version of the game, without the multiple iteration feature.  This is more playable.

If you wish to have the computer play by itself, simply set the variable "ai" to False.
In order to force a jackpot win, simply set the variable "forced_win" to True.

It's really simple to play:  Just press SPACE in order to roll the die (and thus, continue the game), or press Q to bailout (presuming you have at least one strike).

Again, this is just for fun.  No real money nor prizes are awarded.